requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
autoflake==2.3.1
Automat==24.8.1
autopep8==2.3.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-redis==5.4.0
django-resized==1.0.3
django-silk==5.2.0
django-storages==1.14.4
django-templated-mail==1.1.1
django-timezone-field==7.0
django-unfold==0.40.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
flower==2.0.1
gprof2dot==2024.6.6
gunicorn==23.0.0
h11==0.14.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pygraphviz==1.14; platform_system == 'Linux'
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1
tornado==6.4.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
whitenoise==6.8.2
wsproto==1.2.0
zope.interface==7.1.1
